first pass at converting to fixed point...at this point I have gone through every file and converted to fixed point, but no testing has occurred....there were a couple of places I thought I should have converted slightly differently.....I will go do that now   I just wanted to have something checked in.

+++ src/simpull/IForce.java
@@ -31,6 +31,6 @@ package simpull;
 	
 public interface IForce {
 
-	Vector2f getValue(float invMass);
+	Vector2f getValue(int invMass);
 	
 }
